Move voxelize_binary_mask to DataSetFilters
