+++ pyvista/core/filters/poly_data.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
-from typing import Callable
 from typing import cast
 import warnings
 
@@ -4215,480 +4214,6 @@ class PolyDataFilters(DataSetFilters):
         _update_alg(alg, progress_bar, 'Generating Protein Ribbons')
         return _get_output(alg)
 
-    def voxelize_binary_mask(  # type: ignore[misc]
-        self: PolyData,
-        *,
-        background_value: int | float = 0,  # noqa: PYI041
-        foreground_value: int | float = 1,  # noqa: PYI041
-        reference_volume: pyvista.ImageData | None = None,
-        dimensions: VectorLike[int] | None = None,
-        spacing: float | VectorLike[float] | None = None,
-        rounding_func: Callable[[VectorLike[float]], VectorLike[int]] | None = None,
-        cell_length_percentile: float | None = None,
-        cell_length_sample_size: int | None = None,
-        mesh_length_fraction: float | None = None,
-        progress_bar: bool = False,
-    ):
-        """Voxelize :class:`~pyvista.PolyData` as a binary :class:`~pyvista.ImageData` mask.
-
-        The binary mask is a point data array where points inside and outside of the
-        input surface are labelled with ``foreground_value`` and ``background_value``,
-        respectively.
-
-        This filter implements `vtkPolyDataToImageStencil
-        <https://vtk.org/doc/nightly/html/classvtkPolyDataToImageStencil.html>`_. This
-        algorithm operates as follows:
-
-        * The algorithm iterates through the z-slice of the ``reference_volume``.
-        * For each slice, it cuts the input :class:`~pyvista.PolyData` surface to create
-          2D polylines at that z position. It attempts to close any open polylines.
-        * For each x position along the polylines, the corresponding y positions are
-          determined.
-        * For each slice, the grid points are labelled as foreground or background based
-          on their xy coordinates.
-
-        The voxelization can be controlled in several ways:
-
-        #. Specify the output geometry using a ``reference_volume``.
-
-        #. Specify the ``spacing`` explicitly.
-
-        #. Specify the ``dimensions`` explicitly.
-
-        #. Specify the ``cell_length_percentile``. The spacing is estimated from the
-           surface's cells using the specified percentile.
-
-        #. Specify ``mesh_length_fraction``. The spacing is computed as a fraction of
-           the mesh's diagonal length.
-
-        Use ``reference_volume`` for full control of the output mask's geometry. For
-        all other options, the geometry is implicitly defined such that the generated
-        mask fits the bounds of the input surface.
-
-        If no inputs are provided, ``cell_length_percentile=0.1`` (10th percentile) is
-        used by default to estimate the spacing. On systems with VTK < 9.2, the default
-        spacing is set to ``mesh_length_fraction=1/100``.
-
-        .. versionadded:: 0.45.0
-
-        .. note::
-            For best results, ensure the input surface is a closed surface. The
-            surface is considered closed if it has zero :attr:`~pyvista.PolyData.n_open_edges`.
-
-        .. note::
-            This filter returns voxels represented as point data, not :attr:`~pyvista.CellType.VOXEL` cells.
-            This differs from :func:`~pyvista.voxelize` and :func:`~pyvista.voxelize_volume`
-            which return meshes with voxel cells. See :ref:`image_representations_example`
-            for examples demonstrating the difference.
-
-        .. note::
-            This filter does not discard internal surfaces, due, for instance, to
-            intersecting meshes. Instead, the intersection will be considered as
-            background which may produce unexpected results. See `Examples`.
-
-        Parameters
-        ----------
-        background_value : int, default: 0
-            Background value of the generated mask.
-
-        foreground_value : int, default: 1
-            Foreground value of the generated mask.
-
-        reference_volume : pyvista.ImageData, optional
-            Volume to use as a reference. The output will have the same ``dimensions``,
-            ``origin``, ``spacing``, and ``direction_matrix`` as the reference.
-
-        dimensions : VectorLike[int], optional
-            Dimensions of the generated mask image. Set this value to control the
-            dimensions explicitly. If unset, the dimensions are defined implicitly
-            through other parameter. See summary and examples for details.
-
-        spacing : VectorLike[float], optional
-            Approximate spacing to use for the generated mask image. Set this value
-            to control the spacing explicitly. If unset, the spacing is defined
-            implicitly through other parameters. See summary and examples for details.
-
-        rounding_func : Callable[VectorLike[float], VectorLike[int]], optional
-            Control how the dimensions are rounded to integers based on the provided or
-            calculated ``spacing``. Should accept a length-3 vector containing the
-            dimension values along the three directions and return a length-3 vector.
-            :func:`numpy.round` is used by default.
-
-            Rounding the dimensions implies rounding the actual spacing.
-
-            Has no effect if ``reference_volume`` or ``dimensions`` are specified.
-
-        cell_length_percentile : float, optional
-            Cell length percentage ``p`` to use for computing the default ``spacing``.
-            Default is ``0.1`` (10th percentile) and must be between ``0`` and ``1``.
-            The ``p``-th percentile is computed from the cumulative distribution function
-            (CDF) of lengths which are representative of the cell length scales present
-            in the input. The CDF is computed by:
-
-            #. Triangulating the input cells.
-            #. Sampling a subset of up to ``cell_length_sample_size`` cells.
-            #. Computing the distance between two random points in each cell.
-            #. Inserting the distance into an ordered set to create the CDF.
-
-            Has no effect if ``dimension`` or ``reference_volume`` are specified.
-
-            .. note::
-                This option is only available for VTK 9.2 or greater.
-
-        cell_length_sample_size : int, optional
-            Number of samples to use for the cumulative distribution function (CDF)
-            when using the ``cell_length_percentile`` option. ``100 000`` samples are
-            used by default.
-
-        mesh_length_fraction : float, optional
-            Fraction of the surface mesh's length to use for computing the default
-            ``spacing``. Set this to any fractional value (e.g. ``1/100``) to enable
-            this option. This is used as an alternative to using ``cell_length_percentile``.
-
-        progress_bar : bool, default: False
-            Display a progress bar to indicate progress.
-
-        Returns
-        -------
-        pyvista.ImageData
-            Generated binary mask with a ``'mask'``  point data array. The data array
-            has dtype :class:`numpy.uint8` if the foreground and background values are
-            unsigned and less than 256.
-
-        See Also
-        --------
-        pyvista.voxelize
-            Similar function that returns a :class:`~pyvista.UnstructuredGrid` of
-            :attr:`~pyvista.CellType.VOXEL` cells.
-
-        pyvista.voxelize_volume
-            Similar function that returns a :class:`~pyvista.RectilinearGrid` with cell data.
-
-        pyvista.ImageDataFilters.contour_labels
-            Filter that generates surface contours from labeled image data. Can be
-            loosely considered as an inverse of this filter.
-
-        pyvista.ImageDataFilters.points_to_cells
-            Convert voxels represented as points to :attr:`~pyvista.CellType.VOXEL`
-            cells.
-
-        pyvista.ImageData
-            Class used to build custom ``reference_volume``.
-
-        Examples
-        --------
-        Generate a binary mask from a coarse mesh.
-
-        >>> import numpy as np
-        >>> import pyvista as pv
-        >>> from pyvista import examples
-        >>> poly = examples.download_bunny_coarse()
-        >>> mask = poly.voxelize_binary_mask()
-
-        The mask is stored as :class:`~pyvista.ImageData` with point data scalars
-        (zeros for background, ones for foreground).
-
-        >>> mask
-        ImageData (...)
-          N Cells:      7056
-          N Points:     8228
-          X Bounds:     -1.245e-01, 1.731e-01
-          Y Bounds:     -1.135e-01, 1.807e-01
-          Z Bounds:     -1.359e-01, 9.140e-02
-          Dimensions:   22, 22, 17
-          Spacing:      1.417e-02, 1.401e-02, 1.421e-02
-          N Arrays:     1
-
-        >>> np.unique(mask.point_data['mask'])
-        pyvista_ndarray([0, 1], dtype=uint8)
-
-        To visualize it as voxel cells, use :meth:`~pyvista.ImageDataFilters.points_to_cells`,
-        then use :meth:`~pyvista.DataSetFilters.threshold` to extract the foreground.
-
-        We also plot the voxel cells in blue and the input poly data in green for
-        comparison.
-
-        >>> def mask_and_polydata_plotter(mask, poly):
-        ...     voxel_cells = mask.points_to_cells().threshold(0.5)
-        ...
-        ...     plot = pv.Plotter()
-        ...     _ = plot.add_mesh(voxel_cells, color='blue')
-        ...     _ = plot.add_mesh(poly, color='lime')
-        ...     plot.camera_position = 'xy'
-        ...     return plot
-
-        >>> plot = mask_and_polydata_plotter(mask, poly)
-        >>> plot.show()
-
-        The spacing of the mask image is automatically adjusted to match the
-        density of the input.
-
-        Repeat the previous example with a finer mesh.
-
-        >>> poly = examples.download_bunny()
-        >>> mask = poly.voxelize_binary_mask()
-        >>> plot = mask_and_polydata_plotter(mask, poly)
-        >>> plot.show()
-
-        Control the spacing manually instead. Here, a very coarse spacing is used.
-
-        >>> mask = poly.voxelize_binary_mask(spacing=(0.01, 0.04, 0.02))
-        >>> plot = mask_and_polydata_plotter(mask, poly)
-        >>> plot.show()
-
-        Note that the spacing is only approximate. Check the mask's actual spacing.
-
-        >>> mask.spacing
-        (0.009731187485158443, 0.03858340159058571, 0.020112216472625732)
-
-        The actual values may be greater or less than the specified values. Use
-        ``rounding_func=np.floor`` to force all values to be greater.
-
-        >>> mask = poly.voxelize_binary_mask(
-        ...     spacing=(0.01, 0.04, 0.02), rounding_func=np.floor
-        ... )
-        >>> mask.spacing
-        (0.01037993331750234, 0.05144453545411428, 0.020112216472625732)
-
-        Set the dimensions instead of the spacing.
-
-        >>> mask = poly.voxelize_binary_mask(dimensions=(10, 20, 30))
-        >>> plot = mask_and_polydata_plotter(mask, poly)
-        >>> plot.show()
-
-        >>> mask.dimensions
-        (10, 20, 30)
-
-        Create a mask using a reference volume. First generate polydata from
-        an existing mask.
-
-        >>> volume = examples.load_frog_tissues()
-        >>> poly = volume.contour_labels()
-
-        Now create the mask from the polydata using the volume as a reference.
-
-        >>> mask = poly.voxelize_binary_mask(reference_volume=volume)
-        >>> plot = mask_and_polydata_plotter(mask, poly)
-        >>> plot.show()
-
-        Visualize the effect of internal surfaces.
-
-        >>> mesh = pv.Cylinder() + pv.Cylinder((0, 0.75, 0))
-        >>> binary_mask = mesh.voxelize_binary_mask(
-        ...     dimensions=(1, 100, 50)
-        ... ).points_to_cells()
-        >>> plot = pv.Plotter()
-        >>> _ = plot.add_mesh(binary_mask)
-        >>> _ = plot.add_mesh(mesh.slice(), color='red')
-        >>> plot.show(cpos='yz')
-
-        Note how the intersection is excluded from the mask.
-        To include the voxels delimited by internal surfaces in the foreground, the internal
-        surfaces should be removed, for instance by applying a boolean union. Note that
-        this operation in unreliable in VTK but may be performed with external tools such
-        as `vtkbool <https://github.com/zippy84/vtkbool>`_.
-
-        Alternatively, the intersecting parts of the mesh can be processed sequentially.
-
-        >>> cylinder_1 = pv.Cylinder()
-        >>> cylinder_2 = pv.Cylinder((0, 0.75, 0))
-
-        >>> reference_volume = pv.ImageData(
-        ...     dimensions=(1, 100, 50),
-        ...     spacing=(1, 0.0175, 0.02),
-        ...     origin=(0, -0.5 + 0.0175 / 2, -0.5 + 0.02 / 2),
-        ... )
-
-        >>> binary_mask_1 = cylinder_1.voxelize_binary_mask(
-        ...     reference_volume=reference_volume
-        ... ).points_to_cells()
-        >>> binary_mask_2 = cylinder_2.voxelize_binary_mask(
-        ...     reference_volume=reference_volume
-        ... ).points_to_cells()
-
-        >>> binary_mask_1['mask'] = binary_mask_1['mask'] | binary_mask_2['mask']
-
-        >>> plot = pv.Plotter()
-        >>> _ = plot.add_mesh(binary_mask_1)
-        >>> _ = plot.add_mesh(cylinder_1.slice(), color='red')
-        >>> _ = plot.add_mesh(cylinder_2.slice(), color='red')
-        >>> plot.show(cpos='yz')
-
-        When multiple internal surfaces are nested, they are successively treated as
-        interfaces between background and foreground.
-
-        >>> mesh = pv.Tube(radius=2) + pv.Tube(radius=3) + pv.Tube(radius=4)
-        >>> binary_mask = mesh.voxelize_binary_mask(
-        ...     dimensions=(1, 50, 50)
-        ... ).points_to_cells()
-        >>> plot = pv.Plotter()
-        >>> _ = plot.add_mesh(binary_mask)
-        >>> _ = plot.add_mesh(mesh.slice(), color='red')
-        >>> plot.show(cpos='yz')
-
-        """
-        _validation.check_greater_than(self.n_points, 1, name='n_points')
-        _validation.check_greater_than(self.n_cells, 1, name='n_cells')
-
-        def _preprocess_polydata(poly_in):
-            return poly_in.compute_normals().triangulate()
-
-        if reference_volume is not None:
-            if (
-                dimensions is not None
-                or spacing is not None
-                or rounding_func is not None
-                or cell_length_percentile is not None
-                or cell_length_sample_size is not None
-                or mesh_length_fraction is not None
-            ):
-                raise TypeError(
-                    'Cannot specify a reference volume with other geometry parameters. `reference_volume` must define the geometry exclusively.'
-                )
-            _validation.check_instance(reference_volume, pyvista.ImageData, name='reference volume')
-            # The image stencil filters do not support orientation, so we apply the
-            # inverse direction matrix to "remove" orientation from the polydata
-            poly_ijk = self.transform(reference_volume.direction_matrix.T, inplace=False)
-            poly_ijk = _preprocess_polydata(poly_ijk)
-        else:
-            # Compute reference volume geometry
-            if spacing is not None and dimensions is not None:
-                raise TypeError('Spacing and dimensions cannot both be set. Set one or the other.')
-
-            # Need to preprocess so that we have a triangle mesh for computing
-            # cell length percentile
-            poly_ijk = _preprocess_polydata(self)
-
-            if spacing is None:
-                if cell_length_percentile is not None and mesh_length_fraction is not None:
-                    raise TypeError(
-                        'Cell length percentile and mesh length fraction cannot both be set. Set one or the other.'
-                    )
-
-                less_than_vtk92 = pyvista.vtk_version_info < (9, 2)
-                if (
-                    cell_length_percentile is not None or cell_length_sample_size is not None
-                ) and less_than_vtk92:
-                    raise TypeError(
-                        'Cell length percentile and sample size requires VTK 9.2 or greater.'
-                    )
-
-                if mesh_length_fraction is not None or less_than_vtk92:
-                    # Compute spacing from mesh length
-                    mesh_length_fraction = (
-                        1 / 100
-                        if mesh_length_fraction is None
-                        else _validation.validate_number(
-                            mesh_length_fraction,
-                            must_have_dtype=float,
-                            must_be_in_range=[0.0, 1.0],
-                        )
-                    )
-                    spacing = self.length * mesh_length_fraction
-                else:
-                    # Estimate spacing from cell length percentile
-                    cell_length_percentile = (
-                        0.1 if cell_length_percentile is None else cell_length_percentile
-                    )
-                    cell_length_sample_size = (
-                        100_000 if cell_length_sample_size is None else cell_length_sample_size
-                    )
-                    spacing = _length_distribution_percentile(
-                        poly_ijk,
-                        cell_length_percentile,
-                        cell_length_sample_size,
-                        progress_bar=progress_bar,
-                    )
-            else:
-                # Spacing is specified directly. Make sure other params are not set.
-                if cell_length_percentile is not None:
-                    raise TypeError(
-                        'Spacing and cell length percentile cannot both be set. Set one or the other.'
-                    )
-                if mesh_length_fraction is not None:
-                    raise TypeError(
-                        'Spacing and mesh length fraction cannot both be set. Set one or the other.'
-                    )
-
-            # Get initial spacing (will be adjusted later)
-            initial_spacing = _validation.validate_array3(spacing, broadcast=True)
-
-            # Get size of poly data for computing dimensions
-            bnds = self.bounds
-            x_size = bnds.x_max - bnds.x_min
-            y_size = bnds.y_max - bnds.y_min
-            z_size = bnds.z_max - bnds.z_min
-            sizes = np.array((x_size, y_size, z_size))
-
-            if dimensions is None:
-                rounding_func = np.round if rounding_func is None else rounding_func
-                dimensions = np.array(rounding_func(sizes / initial_spacing), dtype=int)
-            elif rounding_func is not None:
-                raise TypeError(
-                    'Rounding func cannot be set when dimensions is specified. Set one or the other.'
-                )
-
-            reference_volume = pyvista.ImageData()
-            reference_volume.dimensions = dimensions  # type: ignore[assignment]
-            # Dimensions are now fixed, now adjust spacing to match poly data bounds
-            # Since we are dealing with voxels as points, we want the bounds of the
-            # points to be 1/2 spacing width smaller than the polydata bounds
-            final_spacing = sizes / np.array(reference_volume.dimensions)
-            reference_volume.spacing = final_spacing
-            reference_volume.origin = np.array(self.bounds[::2]) + final_spacing / 2
-
-        # Init output structure. The image stencil filters do not support
-        # orientation, so we do not set the direction matrix
-        binary_mask = pyvista.ImageData()
-        binary_mask.dimensions = reference_volume.dimensions
-        binary_mask.spacing = reference_volume.spacing
-        binary_mask.origin = reference_volume.origin
-
-        # Init output scalars. Use uint8 dtype if possible.
-        scalars_shape = (binary_mask.n_points,)
-        scalars_dtype: type[np.uint8 | float | int]
-        if all(
-            isinstance(val, int) and val < 256 and val >= 0
-            for val in (background_value, foreground_value)
-        ):
-            scalars_dtype = np.uint8
-        elif all(round(val) == val for val in (background_value, foreground_value)):
-            scalars_dtype = np.int_
-        else:
-            scalars_dtype = np.float64
-        scalars = (  # Init with background value
-            np.zeros(scalars_shape, dtype=scalars_dtype)
-            if background_value == 0
-            else np.ones(scalars_shape, dtype=scalars_dtype) * background_value
-        )
-        binary_mask['mask'] = scalars  # type: ignore[assignment]
-
-        # Make sure that we have a clean triangle-strip polydata
-        # Note: Poly was partially pre-processed earlier
-        poly_ijk = poly_ijk.strip()
-
-        # Convert polydata to stencil
-        poly_to_stencil = _vtk.vtkPolyDataToImageStencil()
-        poly_to_stencil.SetInputData(poly_ijk)
-        poly_to_stencil.SetOutputSpacing(*reference_volume.spacing)
-        poly_to_stencil.SetOutputOrigin(*reference_volume.origin)  # type: ignore[call-overload]
-        poly_to_stencil.SetOutputWholeExtent(*reference_volume.extent)
-        _update_alg(poly_to_stencil, progress_bar, 'Converting polydata')
-
-        # Convert stencil to image
-        stencil = _vtk.vtkImageStencil()
-        stencil.SetInputData(binary_mask)
-        stencil.SetStencilConnection(poly_to_stencil.GetOutputPort())
-        stencil.ReverseStencilOn()
-        stencil.SetBackgroundValue(foreground_value)
-        _update_alg(stencil, progress_bar, 'Generating binary mask')
-        output_volume = _get_output(stencil)
-
-        # Set the orientation of the output
-        output_volume.direction_matrix = reference_volume.direction_matrix
-
-        return pyvista.wrap(output_volume)
-
     def ruled_surface(
         self, *, resolution: VectorLike[int] | None = None, progress_bar: bool = False
     ):
@@ -4750,14 +4275,3 @@ class PolyDataFilters(DataSetFilters):
             alg.SetResolution(resolution)  # type: ignore[arg-type]
         _update_alg(alg, progress_bar, 'Generating ruled surface')
         return _get_output(alg)
-
-
-def _length_distribution_percentile(poly, percentile, cell_length_sample_size, progress_bar):
-    percentile = _validation.validate_number(
-        percentile, must_be_in_range=[0.0, 1.0], name='percentile'
-    )
-    distribution = _vtk.vtkLengthDistribution()
-    distribution.SetInputData(poly)
-    distribution.SetSampleSize(cell_length_sample_size)
-    _update_alg(distribution, progress_bar, 'Computing cell length distribution')
-    return distribution.GetLengthQuantile(percentile)
