+++ tests/doc/doc_image_cache/pyvista-DataSetFilters-voxelize_binary_mask-1_00_00.jpg
